+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 ---
 
+## [2.1.2] - 2019-03-11
+
+### Fixed
+1. Original selected option shown selected in dropdown on page refresh or navigating to page
+
+
+## [2.1.1] - 2018-04-28
+
+### Fixed
+1. Invalid _id_ attribute assigned when a DOM element passed
+
 
 ## [2.1.0] - 2017-10-21